fix: moved some code and removed changed try_into into copy from slice
