+++ crates/crypto/keystore/src/encrypted_keystore.rs
@@ -3,8 +3,11 @@ use std::{fs, path::Path};
 use anyhow::Result;
 use ream_bls::PubKey;
 use serde::{Deserialize, Serialize};
+use sha2::{Digest, Sha256};
 
 use crate::hex_serde;
+use crate::pbkdf2::pbkdf2;
+use crate::scrypt::scrypt;
 
 #[derive(Serialize, Deserialize, Debug, PartialEq)]
 pub struct EncryptedKeystore {
@@ -24,6 +27,22 @@ impl EncryptedKeystore {
     pub fn save_to_file<P: AsRef<Path>>(&self, path: P) -> Result<()> {
         fs::write(path, serde_json::to_string(self)?)?;
         Ok(())
+    }
+
+    pub fn validate_password(&self, password: &[u8]) -> anyhow::Result<bool> {
+        let derived_key = match &self.crypto.kdf.params {
+            KdfParams::Pbkdf2 { c, dklen, prf: _, salt } => {
+                pbkdf2(password, salt, *c, *dklen)?
+            }
+            KdfParams::Scrypt { n, p, r, dklen, salt } => {
+                scrypt(password, salt, *n, *p, *r, *dklen)?
+            }
+        };
+        let derived_key_slice = &derived_key[16..32];
+        let pre_image = [derived_key_slice, &self.crypto.cipher.message].concat();
+        let checksum = Sha256::digest(&pre_image);
+        let valid_password = checksum.as_slice() == self.crypto.checksum.message.as_slice();
+        Ok(valid_password)
     }
 }
 
@@ -178,5 +197,29 @@ mod tests {
                 panic!("Could not load the keystore: {err:?}");
             }
         }
+    }
+
+    #[test]
+    fn password_validation_pbkdf2() {
+        let keystore = EncryptedKeystore::load_from_file("./assets/Pbkdf2TestKeystore.json").unwrap();
+        let password = hex!("7465737470617373776f7264f09f9491");
+
+        assert_eq!(true,keystore.validate_password(&password).unwrap());
+    }
+
+    #[test]
+    fn password_validation_scrypt() {
+        let keystore = EncryptedKeystore::load_from_file("./assets/ScryptKeystore.json").unwrap();
+        let password = b"password123";
+
+        assert_eq!(true,keystore.validate_password(password).unwrap());
+    }
+
+    #[test]
+    fn password_validation_pbkdf2_invalid() {
+        let keystore = EncryptedKeystore::load_from_file("./assets/Pbkdf2TestKeystore.json").unwrap();
+        let password = b"password123";
+
+        assert_eq!(false,keystore.validate_password(password).unwrap());
     }
 }
