+++ crates/crypto/keystore/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod decrypt;
 pub mod encrypted_keystore;
 pub mod hex_serde;
 pub mod hmac;
