+++ crates/crypto/keystore/src/scrypt.rs
@@ -1,7 +1,5 @@
-use std::{convert::TryInto, thread};
-
+use std::thread;
 use anyhow::ensure;
-
 use crate::{pbkdf2::pbkdf2, salsa::salsa20_8_core};
 
 fn scrypt_block_mix(block_size: usize, input_output: &mut [u8]) {
@@ -42,9 +40,9 @@ fn scrypt_romix(block_size: usize, input_output: &mut [u8], cost_parameter: usiz
     let mut temp_block = vec![0u8; block_size_in_bytes];
     for _ in 0..cost_parameter {
         let memory_index = {
-            let last_eight_bytes =
-                &input_output[block_size_in_bytes - 64..block_size_in_bytes - 56];
-            u64::from_le_bytes(last_eight_bytes.try_into().unwrap()) as usize & (cost_parameter - 1)
+            let mut last_eight_bytes_array = [0u8; 8];
+            last_eight_bytes_array.copy_from_slice(&input_output[block_size_in_bytes - 64..block_size_in_bytes - 56]);
+            usize::from_le_bytes(last_eight_bytes_array) & (cost_parameter - 1)
         };
 
         let memory_block = &memory_space
